requirements.txt
streamlit==1.28.1
langchain==0.0.346
langchain-community==0.0.6
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0